更新主机转移逻辑，更新主机转移至模块api
