+++ src/scene_server/host_server/host_service/actions/hosts/relation.go
@@ -352,14 +352,13 @@ func (m *hostModuleConfigAction) moveHostToModuleByName(req *restful.Request, re
 			return http.StatusInternalServerError, nil, defErr.Error(common.CCErrCommJSONUnmarshalFailed)
 		}
 
-		//fmt.Println(moduleURL)
 		conds := make(map[string]interface{})
 		if common.DefaultResModuleName == moduleName {
-			//空闲机
+			//idle module
 			conds[common.BKDefaultField] = common.DefaultResModuleFlag
 			conds[common.BKModuleNameField] = common.DefaultResModuleName
 		} else {
-			//故障机器
+			//fault module
 			conds[common.BKDefaultField] = common.DefaultFaultModuleFlag
 			conds[common.BKModuleNameField] = common.DefaultFaultModuleName
 		}
@@ -385,7 +384,7 @@ func (m *hostModuleConfigAction) moveHostToModuleByName(req *restful.Request, re
 				return http.StatusInternalServerError, nil, defErr.Errorf(common.CCErrHostNotINAPPFail, hostID)
 			}
 			if false == bl {
-				blog.Error("Host does not belong to the current application; error, params:{appid:%d, hostid:%s}", data.ApplicationID, hostID)
+				blog.Error("host do not belong to the current application; error, params:{appid:%d, hostid:%s}", data.ApplicationID, hostID)
 				return http.StatusInternalServerError, nil, defErr.Errorf(common.CCErrHostNotINAPP, hostID)
 			}
